Final commit for HW7. Static field moved to firs position, added getName() and package names changed.
