+++ src/main/java/homework7/goodsolution/Cat.java
@@ -1,19 +1,17 @@
-package homework7.goodDecision;
+package homework7.goodsolution;
 
 public class Cat extends Animal {
 
-    private String name;
     private static int catCount = 0;
 
     public Cat(String name) {
         super(name, 200, 0);
-        this.name = name;
         catCount++;
     }
 
     @Override
     public void swim(int distance) {
-      System.out.println(name + " can't swim.");
+        System.out.println(super.getName() + " can't swim.");
     }
 
     public static int getCatCount() {
